requirements.txt
python-telegram-bot[job-queue]==22.6
requests
apscheduler